requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
Django==2.0.5
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-environ==0.4.4
django-jsonview==1.1.0
future==0.16.0
kombu==4.1.0
lazy-object-proxy==1.3.1
pbr==4.0.2
Pillow==5.1.0
python-telegram-bot==10.1.0
pytz==2018.4
redis==2.10.6
six==1.11.0
tendo==0.2.12
uuid==1.30
uWSGI==2.0.17
vine==1.1.4
